restructured vision tracking
