+++ .gitmodules
@@ -1,4 +1,9 @@
 [submodule "wml"]
 	path = wml
-	url = https://github.com/wml-frc/WML.git
+	url = https://github.com/wml-frc/WML
 	branch = CJ-DevBranch
+
+[submodule "CJ-Vision"]
+	path = CJ-Vision
+	url = https://github.com/wml-frc/CJ-Vision
+	branch = master
