+++ 4788/src/coprocessor/build.gradle
@@ -1,0 +1,1 @@
+apply from: rootProject.file('CJ-Vision/vision.gradle')
