+++ 4788/src/coprocessor/src/main/include/vision.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "VisTrack.h"
+
+class curtin_frc_vision {
+ public:
+  void run();
+
+ private:
+  CJ::VisionTracking vision;
+};
