+++ 4788/src/coprocessor/src/test/cpp/main.cpp
@@ -1,0 +1,8 @@
+#include "gtest/gtest.h"
+
+#include <iostream>
+
+int main(int argc, char** argv) {
+  ::testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
