+++ settings.gradle
@@ -25,4 +25,4 @@ pluginManagement {
 }
 
 rootProject.name = 'CurtinFRC2020'
-include 'common', 'wml', '4788', 'Testbot', 'vision'
+include 'common', 'wml', '4788', 'Testbot', '4788:src:coprocessor'
