--- vision/src/main/include/vision.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "VisTrack.h"
-
-class curtin_frc_vision {
- public:
-  void run();
-
- private:
-  wml::VisionTracking vision;
-};
