--- vision/src/main/resources/system/sshd_config
@@ -1,13 +1,0 @@
-ChallengeResponseAuthentication no
-UsePAM yes
-
-AllowAgentForwarding yes
-AllowTcpForwarding yes
-X11Forwarding yes
-X11DisplayOffset 10
-X11UseLocalhost no
-PrintMotd no
-
-AcceptEnv LANG LC_*
-
-Subsystem	sftp	/usr/lib/openssh/sftp-server
