--- vision/src/main/resources/system/vision.service
@@ -1,11 +1,0 @@
-[Unit]
-Description=Curtin FRC Vision Service
-
-[Service]
-Type=simple
-Restart=always
-ExecStart=/home/vision/curtinFrcVision
-WorkingDirectory=/home/vision
-
-[Install]
-WantedBy=multi-user.target
